Implement bin for compute_shuffled_index
